d07_利用DCGAN生成假脸/requirements-face-api.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.26.0
facenet-pytorch>=2.6.0
torch>=2.0.0
torchvision>=0.15.0
